refactor: migrate to server components and improve security
- Converted client-side pages (dropbox-callback, welcome, portfolio) to server components with proper async handling
- Replaced client-side environment checks with server-side NEXT_PUBLIC_APP_DEV_TOOLS for security
- Added proper TypeScript types for page props with async handling
- Enhanced security by moving sensitive logic to server components
- Improved performance by reducing client-side JavaScript and enabling better caching
- Added proper error boundaries and suspense fallbacks for better UX
